Add filter conditions (names still preliminary) for the smart snippets model https://coveord.atlassian.net/browse/MLX-86
