+++ src/resources/MachineLearning/CaseClassificationConfiguration/CaseClassificationConfigurationInterfaces.ts
@@ -1,27 +1,6 @@
-export enum Operator {
-    Equals = 'EQUALS',
-    NotEquals = 'NOT_EQUALS',
-}
+import {FilterConditions} from '../FilterConditions';
 
 export type CaseClassificationDocumentRequirementStatus = 'OK' | 'INSUFFICIENT_DOCUMENTS';
-
-export interface FilterConditions {
-    /**
-     * The name of the field.
-     * Example: status
-     */
-    field: string;
-    /**
-     * The operator to use to evaluate the condition.
-     * Must be EQUALS or NOT_EQUALS
-     */
-    operator: Operator;
-    /**
-     * The value to use for the condition.
-     * Example: completed
-     */
-    value: string;
-}
 
 export interface DateField {
     /**
