+++ src/resources/MachineLearning/index.ts
@@ -1,5 +1,6 @@
 export * from './MachineLearning';
 export * from './MachineLearningInterfaces';
+export * from './FilterConditions';
 export * from './Models/';
 export * from './ModelInformation/';
 export * from './ModelConfiguration/';
